requirements.txt
streamlit
pillow
python-dotenv
PyPDF2
pdf2image